images almost
darcs-hash:20060630093105-55968-eed8efe246bc0ca203bd91ad1ab005de7f69f942.gz

+++ src/base/post_set.cpp
@@ -141,7 +141,7 @@ Uint32 PostSet::min_article_no(void)
     return article_no;
 }
 
-Uint64 PostSet::num_bytes(void)
+unsigned long long int PostSet::num_bytes(void)
 {
     if(!has_pieces_loaded){
         return _num_bytes;
